bug fix in compile bundle + enable ipex cpu + uplift oneapi version (#5570)
* bug fix in compile bundle

* enable ipex cpu compilation in compile bundle

* update compile bundle remove_directory api usage

* work with torch/vision/audio nightly version separately in compile bundle

* remove llvm compilation

* uplift oneapi version

+++ examples/gpu/llm/tools/compilation_utils.py
@@ -112,7 +112,7 @@ def check_system_commands(commands):
     assert absent == 0, 'Required system command(s) not found.'
     return ret
 
-def remove_directory(directory):
+def _remove_directory(directory):
     if SYSTEM == 'Windows':
         for root, dirs, files in os.walk(directory):
             for dir_name in dirs:
@@ -123,13 +123,14 @@ def remove_directory(directory):
                 os.chmod(file_path, stat.S_IWRITE)
     shutil.rmtree(directory)
 
-def remove_file_dir(item):
-    if os.path.isfile(item):
+def remove_file_dir(item, isfile = False, isdir = False):
+    if not isfile and not isdir:
+        isfile = True
+        isdir = True
+    if os.path.isfile(item) and isfile:
         os.remove(item)
-    elif os.path.isdir(item):
-        remove_directory(item)
-    else:
-        pass
+    if os.path.isdir(item) and isdir:
+        _remove_directory(item)
 
 def clear_directory(directory):
     for item in os.listdir(directory):
@@ -158,6 +159,10 @@ def update_source_code(dir_name,
                   silent = True)
         exec_cmds(f'git checkout {branch}',
                   cwd = dir_target,
+                  show_command = show_command)
+        exec_cmds('git pull',
+                  cwd = dir_target,
+                  exit_on_failure = False,
                   show_command = show_command)
     exec_cmds('''git submodule sync
                  git submodule update --init --recursive''',
